move well-formed unicode strings proposal to stage 4
https://github.com/babel/proposals/issues/88#issuecomment-1548169487
https://github.com/tc39/proposal-is-usv-string/pull/31
