+++ packages/core-js/actual/instance/is-well-formed.js
@@ -1,10 +1,3 @@
-var isPrototypeOf = require('../../internals/object-is-prototype-of');
-var method = require('../string/virtual/is-well-formed');
+var parent = require('../../stable/instance/is-well-formed');
 
-var StringPrototype = String.prototype;
-
-module.exports = function (it) {
-  var own = it.isWellFormed;
-  return typeof it == 'string' || it === StringPrototype
-    || (isPrototypeOf(StringPrototype, it) && own === StringPrototype.isWellFormed) ? method : own;
-};
+module.exports = parent;
