+++ packages/core-js/actual/string/is-well-formed.js
@@ -1,3 +1,6 @@
+// TODO: Remove from `core-js@4`
 require('../../modules/esnext.string.is-well-formed');
 
-module.exports = require('../../internals/entry-unbind')('String', 'isWellFormed');
+var parent = require('../../stable/string/is-well-formed');
+
+module.exports = parent;
