+++ packages/core-js/actual/string/to-well-formed.js
@@ -1,3 +1,6 @@
+// TODO: Remove from `core-js@4`
 require('../../modules/esnext.string.to-well-formed');
 
-module.exports = require('../../internals/entry-unbind')('String', 'toWellFormed');
+var parent = require('../../stable/string/to-well-formed');
+
+module.exports = parent;
