+++ packages/core-js/actual/string/virtual/index.js
@@ -1,5 +1,6 @@
 var parent = require('../../../stable/string/virtual');
 
+// TODO: Remove from `core-js@4`
 require('../../../modules/esnext.string.is-well-formed');
 require('../../../modules/esnext.string.to-well-formed');
