+++ packages/core-js/es/string/index.js
@@ -6,6 +6,7 @@ require('../../modules/es.string.code-point-at');
 require('../../modules/es.string.at-alternative');
 require('../../modules/es.string.ends-with');
 require('../../modules/es.string.includes');
+require('../../modules/es.string.is-well-formed');
 require('../../modules/es.string.match');
 require('../../modules/es.string.match-all');
 require('../../modules/es.string.pad-end');
@@ -17,6 +18,7 @@ require('../../modules/es.string.search');
 require('../../modules/es.string.split');
 require('../../modules/es.string.starts-with');
 require('../../modules/es.string.substr');
+require('../../modules/es.string.to-well-formed');
 require('../../modules/es.string.trim');
 require('../../modules/es.string.trim-start');
 require('../../modules/es.string.trim-end');
