+++ packages/core-js/es/string/virtual/is-well-formed.js
@@ -1,0 +1,3 @@
+require('../../../modules/es.string.is-well-formed');
+
+module.exports = require('../../../internals/entry-virtual')('String').isWellFormed;
