+++ packages/core-js/es/string/virtual/to-well-formed.js
@@ -1,0 +1,3 @@
+require('../../../modules/es.string.to-well-formed');
+
+module.exports = require('../../../internals/entry-virtual')('String').toWellFormed;
