+++ packages/core-js/stable/instance/is-well-formed.js
@@ -1,0 +1,3 @@
+var parent = require('../../es/instance/is-well-formed');
+
+module.exports = parent;
