+++ packages/core-js/stable/instance/to-well-formed.js
@@ -1,0 +1,3 @@
+var parent = require('../../es/instance/to-well-formed');
+
+module.exports = parent;
