+++ packages/core-js/stable/string/is-well-formed.js
@@ -1,0 +1,3 @@
+var parent = require('../../es/string/is-well-formed');
+
+module.exports = parent;
