+++ packages/core-js/stable/string/to-well-formed.js
@@ -1,0 +1,3 @@
+var parent = require('../../es/string/to-well-formed');
+
+module.exports = parent;
