+++ packages/core-js/stable/string/virtual/is-well-formed.js
@@ -1,0 +1,3 @@
+var parent = require('../../../es/string/virtual/is-well-formed');
+
+module.exports = parent;
