+++ packages/core-js/stable/string/virtual/to-well-formed.js
@@ -1,0 +1,3 @@
+var parent = require('../../../es/string/virtual/to-well-formed');
+
+module.exports = parent;
