+++ tests/entries/unit.mjs
@@ -225,6 +225,8 @@ for (PATH of ['core-js-pure', 'core-js']) {
     ok(typeof load(NS, 'string/replace-all') == 'function');
     ok(load(NS, 'string/pad-start')('a', 3) === '  a');
     ok(load(NS, 'string/pad-end')('a', 3) === 'a  ');
+    ok(load(NS, 'string/is-well-formed')('a'));
+    ok(load(NS, 'string/to-well-formed')('a') === 'a');
     ok('next' in load(NS, 'string/iterator')('qwe'));
     ok(load(NS, 'string/virtual/at').call('a', 0) === 'a');
     ok(load(NS, 'string/virtual/code-point-at').call('a', 0) === 97);
@@ -255,6 +257,8 @@ for (PATH of ['core-js-pure', 'core-js']) {
     ok(typeof load(NS, 'string/virtual/sup') == 'function');
     ok(load(NS, 'string/virtual/pad-start').call('a', 3) === '  a');
     ok(load(NS, 'string/virtual/pad-end').call('a', 3) === 'a  ');
+    ok(load(NS, 'string/virtual/is-well-formed').call('a'));
+    ok(load(NS, 'string/virtual/to-well-formed').call('a') === 'a');
     ok('next' in load(NS, 'string/virtual/iterator').call('qwe'));
     ok('padEnd' in load(NS, 'string/virtual'));
     ok('raw' in load(NS, 'string'));
@@ -448,6 +452,12 @@ for (PATH of ['core-js-pure', 'core-js']) {
     ok(typeof instanceKeys([]) == 'function');
     ok(instanceKeys([]).call([1, 2, 3]).next().value === 0);
 
+    const instanceIsWellFormed = load(NS, 'instance/is-well-formed');
+    ok(typeof instanceIsWellFormed == 'function');
+    ok(instanceIsWellFormed({}) === undefined);
+    ok(typeof instanceIsWellFormed('') == 'function');
+    ok(instanceIsWellFormed('').call('a'));
+
     const instanceLastIndexOf = load(NS, 'instance/last-index-of');
     ok(typeof instanceLastIndexOf == 'function');
     ok(instanceLastIndexOf({}) === undefined);
@@ -557,6 +567,12 @@ for (PATH of ['core-js-pure', 'core-js']) {
     ok(instanceToSpliced({}) === undefined);
     ok(typeof instanceToSpliced([]) == 'function');
     ok(instanceToSpliced([]).call([3, 2, 1], 1, 1, 4, 5).length === 4);
+
+    const instanceToWellFormed = load(NS, 'instance/to-well-formed');
+    ok(typeof instanceToWellFormed == 'function');
+    ok(instanceToWellFormed({}) === undefined);
+    ok(typeof instanceToWellFormed('') == 'function');
+    ok(instanceToWellFormed('').call('a') === 'a');
 
     const instanceTrimEnd = load(NS, 'instance/trim-end');
     ok(typeof instanceTrimEnd == 'function');
@@ -682,10 +698,6 @@ for (PATH of ['core-js-pure', 'core-js']) {
     ok(load(NS, 'set/is-superset-of')(new Set([1, 2, 3, 4]), new Set([1, 2, 3])));
     ok(load(NS, 'set/symmetric-difference')(new Set([1, 2, 3]), new Set([3, 4, 5])).size === 4);
     ok(load(NS, 'set/union')(new Set([1, 2, 3]), new Set([3, 4, 5])).size === 5);
-    ok(load(NS, 'string/is-well-formed')('a'));
-    ok(load(NS, 'string/virtual/is-well-formed').call('a'));
-    ok(load(NS, 'string/to-well-formed')('a') === 'a');
-    ok(load(NS, 'string/virtual/to-well-formed').call('a') === 'a');
     ok(load(NS, 'symbol/dispose'));
     ok(load(NS, 'symbol/metadata'));
     ok(new (load(NS, 'suppressed-error'))(1, 2).suppressed === 2);
@@ -716,18 +728,6 @@ for (PATH of ['core-js-pure', 'core-js']) {
     ok(instanceGroupByToMap({}) === undefined);
     ok(typeof instanceGroupByToMap([]) == 'function');
     ok(instanceGroupByToMap([]).call([1, 2, 3], it => it % 2).get(1).length === 2);
-
-    const instanceIsWellFormed = load(NS, 'instance/is-well-formed');
-    ok(typeof instanceIsWellFormed == 'function');
-    ok(instanceIsWellFormed({}) === undefined);
-    ok(typeof instanceIsWellFormed('') == 'function');
-    ok(instanceIsWellFormed('').call('a'));
-
-    const instanceToWellFormed = load(NS, 'instance/to-well-formed');
-    ok(typeof instanceToWellFormed == 'function');
-    ok(instanceToWellFormed({}) === undefined);
-    ok(typeof instanceToWellFormed('') == 'function');
-    ok(instanceToWellFormed('').call('a') === 'a');
   }
 
   for (const NS of ['full', 'features']) {
