+++ tests/unit-global/es.string.is-well-formed.js
